% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_data.R
\name{cache_workshop_data}
\alias{cache_workshop_data}
\alias{get_workshop_data}
\title{Download and cache workshop data from CytoverseBioc2023:::zip_link}
\usage{
cache_workshop_data(force = FALSE)

get_workshop_data(path)
}
\arguments{
\item{force}{Should we download and cache again even if it appears that data have already been downloaded?}

\item{path}{\code{character} giving a path to file(s) in the original tarball}
}
\value{
\code{\link[BiocFileCache:BiocFileCache-class]{BiocFileCache::BiocFileCache()}}, invisibly
}
\description{
Download and cache workshop data from CytoverseBioc2023:::zip_link
}
\section{Functions}{
\itemize{
\item \code{get_workshop_data()}: return a \code{data.frame} with \code{rpath} pointing to the location of files whose paths match \code{path}

}}
