% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop_cs.R
\name{make_cytoset}
\alias{make_cytoset}
\title{Load a set of fcs files as a cytoset and attach mock metadata}
\usage{
make_cytoset(only_TNK = FALSE)
}
\arguments{
\item{only_TNK}{only return the TNK panel}
}
\value{
\code{\link[flowWorkspace:cytoset]{flowWorkspace::cytoset()}}
}
\description{
Load a set of fcs files as a cytoset and attach mock metadata
}
