% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop_gs.R
\name{make_transformed_gs}
\alias{make_transformed_gs}
\title{Transform and attach gates to TNK-CR data}
\usage{
make_transformed_gs()
}
\value{
\code{\link[flowWorkspace:GatingSet-class]{flowWorkspace::GatingSet()}}
}
\description{
Transform and attach gates to TNK-CR data
}
